+++ CHANGELOG.md
@@ -5,3 +5,5 @@
 * The dataset config file now supports 'single' and 'multiple' dataset types
 * New 'single' datasets has been added to the agilehost3-prod5.yml
 * A bug in the photometry code has been spotted and fixed. The bug caused fixed region integrations.
+* Implementation of a custom loss using the Keras API.
+* Upgrade of the 'predictions' plot (plot_predictions_v2).